Code cleanup - Get rid of warnings.

+++ RocketBot2/WebSocketHandler/ActionCommands/TransferPokemonHandler.cs
@@ -1,10 +1,10 @@
 ﻿#region using directives
 
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using PoGo.NecroBot.Logic.State;
 using PoGo.NecroBot.Logic.Tasks;
 using SuperSocket.WebSocket;
-using System.Collections.Generic;
 
 #endregion
 
@@ -21,7 +21,14 @@ namespace RocketBot2.WebSocketHandler.ActionCommands
 
         public async Task Handle(ISession session, WebSocketSession webSocketSession, dynamic message)
         {
-            await TransferPokemonTask.Execute(session, session.CancellationTokenSource.Token, new List<ulong> { message.PokemonId });
+            await TransferPokemonTask.Execute(
+                session,
+                session.CancellationTokenSource.Token,
+                new List<ulong>
+                {
+                    (ulong) message.PokemonId
+                }
+            );
         }
     }
 }
